requirements.txt
ansible==2.10.7
ansible-base==2.10.5
apache-libcloud==3.2.0
appdirs==1.4.3
apturl==0.5.2
argcomplete==1.8.1
asgiref==3.2.3
astroid==2.4.2
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.4
boto==2.49.0
boto3==1.9.238
botocore==1.12.253
Brlapi==0.8.2
brotlipy==0.7.0
CacheControl==0.12.6
cairocffi==1.1.0
CairoSVG==2.4.2
captcha==0.3
certifi==2019.11.28
cffi==1.14.6
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
configparser==4.0.2
contextlib2==0.6.0
cryptography==3.3.2
cssselect2==0.2.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.2
distro==1.4.0
distro-info==1.0
dj-database-url==0.5.0
Django==3.0.3
django-classy-tags==1.0.0
django-cleanup==4.0.0
django-cookie-law==2.0.3
django-cors-headers==3.2.1
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-disqus==0.5
django-environ==0.4.5
django-filter==2.2.0
django-fluent-contents==2.0.7
django-fluent-utils==2.0.1
django-form-designer-ai==1.0.1
django-parler==2.0.1
django-picklefield==0.3.2
django-polymorphic==2.1.2
django-recaptcha==2.0.6
django-restframework==0.0.1
django-tag-parser==3.2
django-template-analyzer==1.6.2
django-tinymce==3.1.0
django-wysiwyg==0.8.0
djangorestframework==3.12.1
dnspython==1.16.0
docutils==0.15.2
duplicity==0.8.17
ecdsa==0.16.1
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
googlevoice==3.0.0
gpg===1.14.0-unknown
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
ipaddr==2.2.0
isort==5.6.4
jeepney==0.6.0
Jinja2==2.11.2
jmespath==0.9.4
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.16.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
monotonic==1.5
msgpack==0.6.2
netaddr==0.7.19
netifaces==0.10.9
ntlm-auth==1.4.0
numpy==1.19.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pandas==1.1.4
paramiko==2.7.2
pep517==0.8.2
pexpect==4.8.0
Pillow==8.1.2
platformdirs==2.2.0
progress==1.5
protobuf==3.12.4
psycopg2-binary==2.8.6
pycairo==1.16.2
pycparser==2.19
pycryptodomex==3.9.7
pycups==2.0.1
PyGObject==3.38.0
PyGoogleVoice==0.5.2
PyJWT==1.7.1
pykerberos==1.1.14
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==2.4.6
Pyphen==0.10.0
pyRFC3339==1.1
pytest-runner==5.2
python-apt==2.1.7+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.39
pytoml==0.1.21
pytz==2019.3
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.3.1
reportlab==3.5.66
requests==2.22.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
retrying==1.3.3
s3transfer==0.2.1
SecretStorage==3.3.1
selinux==3.1
simplejson==3.17.2
six==1.14.0
smsapi-client==2.4.3
soupsieve==2.0
sqlparse==0.3.0
systemd-python==234
tinycss2==1.0.2
toml==0.10.2
ubuntu-advantage-tools==24.4
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.7.1
wadllib==1.3.5
WeasyPrint==52.1
webencodings==0.5.1
whitenoise==4.1.4
wrapt==1.12.1
xdg==5
xkit==0.0.0
xmltodict==0.12.0
